requirements.txt
bottle==0.12.23
ctypes_snappy==1.03
diskcache==5.4.0
huggingface_hub==0.11.1
matplotlib==3.6.2
numpy==1.24.1
opencv_python_headless==4.7.0.68
Pillow==9.4.0
psutil==5.9.4
PyYAML==6.0
requests==2.28.2
tensorflow==2.10.1
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
wand==0.6.11
